Fix reports route
